requirements.txt
robin-stocks>=3.0.0
gspread>=5.0.0
oauth2client>=4.1.3
python-dotenv>=0.19.0
pytz>=2021.3
python-dateutil>=2.8.2
pandas>=1.3.0
cryptography>=3.4.8
pyinstaller>=5.0.0